requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
requests==2.32.5
python-dotenv==1.1.1
pydantic==2.11.7
rich==14.1.0
rank-bm25==0.2.2
nltk==3.9.1
sentence-transformers==5.1.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.43
pgvector==0.4.1
